+++ R/0_imports.R
@@ -8,6 +8,10 @@
 
 # ------------------------------------------------------------------------------
 
+utils::globalVariables(c("x", "y"))
+
+# ------------------------------------------------------------------------------
+
 # The functions below define the model information. These access the model
 # environment inside of parsnip so they have to be executed once parsnip has
 # been loaded.
@@ -16,9 +20,12 @@
   # This defines discrim_linear in the model database
   make_discrim_linear_MASS()
   make_discrim_linear_mda()
+  make_discrim_linear_sda()
+  make_discrim_linear_sparsediscrim()
 
   # This defines discrim_quad in the model database
   make_discrim_quad_MASS()
+  make_discrim_quad_sparsediscrim()
 
   # This defines discrim_regularized in the model database
   make_discrim_regularized()