+++ R/discrim_quad.R
@@ -1,29 +1,33 @@
-#' General Interface for Quadratic Discriminant Models
+#' Quadratic discriminant analysis
 #'
-#' `discrim_quad()` is a way to generate a _specification_ of a quadratic
-#'  discriminant analysis (QDA) model before fitting and allows the model to be
-#'  created using different packages in R.
+#' @description
+#'
+#' `discrim_quad()` defines a model that estimates a multivariate
+#'  distribution for the predictors separately for the data in each class
+#'  (usually Gaussian with separate covariance matrices). Bayes' theorem is used
+#'  to compute the probability of each class, given the predictor values.
+#'
+#' There are different ways to fit this model. See the engine-specific pages
+#' for more details:
+#'
+#' \Sexpr[stage=render,results=rd]{parsnip:::make_engine_list("discrim_quad", "discrim")}
+#'
+#' More information on how \pkg{parsnip} is used for modeling is at
+#' \url{https://www.tidymodels.org/}.
 #'
 #' @param mode A single character string for the type of model. The only
 #'  possible value for this model is "classification".
-#' @details
-#' For `discrim_quad()`, the mode will always be "classification".
+#' @param engine A single character string specifying what computational engine to use
+#'  for fitting.
+#' @param regularization_method A character string for the type of regularized
+#'  estimation. Possible values are: "`diagonal`", "`shrink_cov`", and
+#'  "`shrink_mean`" (`sparsediscrim` engine only).
 #'
-#' The model can be created using the `fit()` function using the following
-#'  _engines_:
-
-#' \itemize{
-#' \item \pkg{R}:  `"MASS"`
-#' }
+#' @template spec-details
 #'
+#' @template spec-references
 #'
-#' This argument is converted to its specific names at the time that the model
-#'  is fit. Other options and argument can be set using `set_engine()`. If left
-#'  to their defaults here (`NULL`), the values are taken from the underlying
-#'  model functions. If parameters need to be modified, `update()` can be used
-#'  in lieu of recreating the object from scratch.
-#'
-#' @includeRmd man/rmd/discrim-quad-engine.Rmd
+#' @seealso \Sexpr[stage=render,results=rd]{parsnip:::make_seealso_list("discrim_quad", "discrim")}
 #'
 #' @examples
 #' parabolic_grid <-
@@ -47,9 +51,9 @@
 #'   coord_equal()
 #' @export
 discrim_quad <-
-  function(mode = "classification") {
+  function(mode = "classification", engine = "MASS", regularization_method = NULL) {
 
-    args <- list()
+    args <- list(regularization_method = rlang::enquo(regularization_method))
 
     parsnip::new_model_spec(
       "discrim_quad",
@@ -57,7 +61,7 @@ discrim_quad <-
       eng_args = NULL,
       mode = mode,
       method = NULL,
-      engine = NULL
+      engine = engine
     )
   }
 
@@ -76,16 +80,16 @@ print.discrim_quad <- function(x, ...) {
 
 # ------------------------------------------------------------------------------
 
-#' @inheritParams update.discrim_flexible
-#' @param object A quadratic discriminant model specification.
 #' @method update discrim_quad
-#' @rdname discrim_quad
+#' @rdname discrim_update
+#' @inheritParams discrim_quad
 #' @export
 update.discrim_quad <-
   function(object,
+           regularization_method = NULL,
            fresh = FALSE, ...) {
     parsnip::update_dot_check(...)
-    args <- list()
+    args <- list(regularization_method = rlang::enquo(regularization_method))
 
     if (fresh) {
       object$args <- args