+++ R/naive_Bayes.R
@@ -1,43 +1,32 @@
-#' General Interface for Naive Bayes Models
+#' Naive Bayes models
 #'
-#' `naive_Bayes()` is a way to generate a _specification_ of a model before
-#'  fitting and allows the model to be created using different packages in R.
+#' @description
 #'
-#' @param mode A single character string for the type of model. The only
-#'  possible value for this model is "classification".
+#' `naive_Bayes()` defines a model uses Bayes' theorem to compute the
+#' probability of each class, given the predictor values.
+#'
+#' There are different ways to fit this model. See the engine-specific pages
+#' for more details:
+#'
+#' \Sexpr[stage=render,results=rd]{parsnip:::make_engine_list("naive_Bayes", "discrim")}
+#'
+#' More information on how \pkg{parsnip} is used for modeling is at
+#' \url{https://www.tidymodels.org/}.
+#'
+#' @inheritParams discrim_linear
 #' @param smoothness An non-negative number representing the the relative
 #'  smoothness of the class boundary. Smaller examples result in model flexible
 #'  boundaries and larger values generate class boundaries that are less
 #'  adaptable
 #' @param Laplace A non-negative value for the Laplace correction to smoothing
 #' low-frequency counts.
-#' @details
 #'
-#' The main
-#'  arguments for the model are:
-#' \itemize{
-#'   \item \code{smoothness}: The total amount of regularization
-#'   in the model. Note that this only used for the "klaR" engine where it is a
-#'   pure L2 smoothness (a.k.a ridge regression).
-#'   \item \code{Laplace}: Laplace correction for smoothing low-frequency counts.
-#' }
-#' These arguments are converted to their specific names at the time that the
-#'  model is fit. Other options and argument can be set using `set_engine()`. If
-#'  left to their defaults here (`NULL`), the values are taken from the
-#'  underlying model functions. If parameters need to be modified, `update()`
-#'  can be used in lieu of recreating the object from scratch.
+#' @template spec-details
 #'
-#' For `naive_Bayes()`, the mode will always be "classification".
+#' @template spec-references
 #'
-#' The model can be created using the `fit()` function using the following _engines_:
-#' \itemize{
-#' \item \pkg{R}:  `"klaR"`(the default) or `"naivebayes"`
-#' }
+#' @seealso \Sexpr[stage=render,results=rd]{parsnip:::make_seealso_list("naive_Bayes", "discrim")}
 #'
-#'
-#' @includeRmd man/rmd/naive-Bayes-engine.Rmd
-#'
-
 #' @examples
 #' parabolic_grid <-
 #'   expand.grid(X1 = seq(-5, 5, length = 100),
@@ -60,7 +49,7 @@
 #'   coord_equal()
 #' @export
 naive_Bayes <-
-  function(mode = "classification", smoothness = NULL, Laplace = NULL) {
+  function(mode = "classification", engine = "klaR", smoothness = NULL, Laplace = NULL) {
     args <-
       list(
         smoothness = rlang::enquo(smoothness),
@@ -73,7 +62,7 @@ naive_Bayes <-
       eng_args = NULL,
       mode = mode,
       method = NULL,
-      engine = NULL
+      engine = engine
     )
   }
 
@@ -92,17 +81,9 @@ print.naive_Bayes <- function(x, ...) {
 
 # ------------------------------------------------------------------------------
 
-#' @inheritParams update.discrim_flexible
-#' @param object A linear discriminant model specification.
-#' @examples
-#'
-#'
-#' model <- naive_Bayes(smoothness = 0.1)
-#' model
-#' update(model, smoothness = 1)
-#' update(model, smoothness = 1, fresh = TRUE)
 #' @method update naive_Bayes
-#' @rdname naive_Bayes
+#' @rdname discrim_update
+#' @inheritParams naive_Bayes
 #' @export
 update.naive_Bayes <-
   function(object,